SSG-6625 add confirmation dialog when removing a rule
git-svn-id: https://thesource.l7tech.com/thesource/UneasyRooster/trunk@34969 219f1c6e-4a0e-43c8-96f2-6e749be13fd4

+++ modules/gateway/console/src/main/java/com/l7tech/console/panels/SsgFirewallManagerDialog.java
@@ -195,6 +195,15 @@ public class SsgFirewallManagerDialog extends JDialog {
             @Override
             public void actionPerformed(final ActionEvent e) {
                 final SsgFirewallRule rule = getSelectedInput();
+                int result = JOptionPane.showConfirmDialog(SsgFirewallManagerDialog.this,
+                        "Are you sure you want to remove the firewall rule \"" + rule.getName() + "\"?",
+                        "Confirm Removal",
+                        JOptionPane.YES_NO_CANCEL_OPTION,
+                        JOptionPane.QUESTION_MESSAGE);
+                if (result != JOptionPane.YES_OPTION)
+                    return;
+
+
                 int rowIndex = firewallRulesTable.getSelectedRow();
                 try {
                     Registry.getDefault().getTransportAdmin().deleteFirewallRule(rule.getOid());
